backend/ai-image-detector-service/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mlxtend==0.23.4
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
fastapi==0.101.0
uvicorn[standard]==0.24.0
firebase-admin==6.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
python-multipart==0.0.9
